requirements.txt
# Data Processing (compatible with Python 3.12)
numpy>=1.21.6,<2.0.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.2
mlflow

# Data Validation
openpyxl>=3.1.5

# Visualization
matplotlib==3.8.4
seaborn>=0.13.2

# Web Requests
requests>=2.32.3

# Feature Store
pyyaml>=6.0.2

# Utilities
pyyaml>=6.0.2
python-dateutil>=2.9.0.post0
tqdm>=4.66.5

# Logging and Monitoring
colorlog>=6.8.2

# File Operations
pathlib2>=2.3.7.post1

# Cloud Storage
boto3>=1.35.39

# Environment and Configuration
python-dotenv>=1.0.1

# Development and Testing
pytest>=8.3.3
pytest-cov>=5.0.0

# Database Support
sqlalchemy>=2.0.0

# Data Versioning
dvc>=3.0.0
dvc-s3>=3.0.0

# Jupyter (for EDA notebooks)
jupyter>=1.0.0
ipykernel>=6.29.5

# Orchestration
apache-airflow>=2.8.0
schedule>=1.2.0